Refactor common/distribution to client
